feat:registerinnhenting for andre ytelser (k9). Splitter engangsstønad til egen pakke, setter resten til default
